backend/main-api-global-mmm/requirements.txt
annotated-types==0.6.0
attrs==23.2.0
bayesian-optimization==1.4.3
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cma==3.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dacite==1.8.1
dill==0.3.8
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Session==0.6.0
fonttools==4.48.1
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==22.0.0
htmlmin==0.1.12
idna==3.6
ImageHash==4.3.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.8.2
multimethod==1.11
multiprocess==0.70.14
nestd==0.3.1
networkx==3.2.1
nevergrad==1.0.1
numba==0.58.1
numpy==1.25.2
packaging==23.2
pandas==1.5.3
patsy==0.5.6
pfzy==0.3.4
phik==0.12.4
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.12.2
six==1.16.0
statsmodels==0.14.1
tangled-up-in-unicode==0.2.0
threadpoolctl==3.2.0
tqdm==4.66.1
typeguard==4.1.5
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
visions==0.7.5
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wordcloud==1.9.3
ydata-profiling==4.6.4
zipp==3.17.0
flask_apscheduler==1.13.1
flask_socketio==5.3.6